# gen/http/lfx_v2_formation_service/server/encode_decode.go
// Code generated by goa v3.25.3, DO NOT EDIT.
//
// lfx_v2_formation_service HTTP server encoders and decoders
//
// Command:
// $ goa gen
// github.com/linuxfoundation/lfx-v2-formation-service/cmd/formation-api/design

package server

import (
	"context"
	"errors"
	"net/http"
	"strconv"
	"strings"

	lfxv2formationservice "github.com/linuxfoundation/lfx-v2-formation-service/gen/lfx_v2_formation_service"
	lfxv2formationserviceviews "github.com/linuxfoundation/lfx-v2-formation-service/gen/lfx_v2_formation_service/views"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeGetFormationResponse returns an encoder for responses returned by the
// lfx_v2_formation_service get_formation endpoint.
func EncodeGetFormationResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res := v.(*lfxv2formationserviceviews.FormationChecklist)
		enc := encoder(ctx, w)
		body := NewGetFormationResponseBody(res.Projected)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGetFormationRequest returns a decoder for requests sent to the
// lfx_v2_formation_service get_formation endpoint.
func DecodeGetFormationRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*lfxv2formationservice.GetFormationPayload, error) {
	return func(r *http.Request) (*lfxv2formationservice.GetFormationPayload, error) {
		var payload *lfxv2formationservice.GetFormationPayload
		var (
			projectUID  string
			bearerToken *string

			params = mux.Vars(r)
		)
		projectUID = params["project_uid"]
		bearerTokenRaw := r.Header.Get("Authorization")
		if bearerTokenRaw != "" {
			bearerToken = &bearerTokenRaw
		}
		payload = NewGetFormationPayload(projectUID, bearerToken)
		if payload.BearerToken != nil {
			if strings.Contains(*payload.BearerToken, " ") {
				// Remove authorization scheme prefix (e.g. "Bearer")
				cred := strings.SplitN(*payload.BearerToken, " ", 2)[1]
				payload.BearerToken = &cred
			}
		}

		return payload, nil
	}
}

// EncodeGetFormationError returns an encoder for errors returned by the
// get_formation lfx_v2_formation_service endpoint.
func EncodeGetFormationError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "NotFound":
			var res *lfxv2formationservice.NotFoundError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewGetFormationNotFoundResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusNotFound)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeGetFormationActivityResponse returns an encoder for responses returned
// by the lfx_v2_formation_service get_formation_activity endpoint.
func EncodeGetFormationActivityResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res := v.(*lfxv2formationserviceviews.FormationActivityPage)
		enc := encoder(ctx, w)
		body := NewGetFormationActivityResponseBody(res.Projected)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGetFormationActivityRequest returns a decoder for requests sent to the
// lfx_v2_formation_service get_formation_activity endpoint.
func DecodeGetFormationActivityRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*lfxv2formationservice.GetFormationActivityPayload, error) {
	return func(r *http.Request) (*lfxv2formationservice.GetFormationActivityPayload, error) {
		var payload *lfxv2formationservice.GetFormationActivityPayload
		var (
			projectUID  string
			cursor      *string
			limit       int
			bearerToken *string
			err         error

			params = mux.Vars(r)
		)
		projectUID = params["project_uid"]
		qp := r.URL.Query()
		cursorRaw := qp.Get("cursor")
		if cursorRaw != "" {
			cursor = &cursorRaw
		}
		{
			limitRaw := qp.Get("limit")
			if limitRaw == "" {
				limit = 20
			} else {
				v, err2 := strconv.ParseInt(limitRaw, 10, strconv.IntSize)
				if err2 != nil {
					err = goa.MergeErrors(err, goa.InvalidFieldTypeError("limit", limitRaw, "integer"))
				}
				limit = int(v)
			}
		}
		bearerTokenRaw := r.Header.Get("Authorization")
		if bearerTokenRaw != "" {
			bearerToken = &bearerTokenRaw
		}
		if err != nil {
			return payload, err
		}
		payload = NewGetFormationActivityPayload(projectUID, cursor, limit, bearerToken)
		if payload.BearerToken != nil {
			if strings.Contains(*payload.BearerToken, " ") {
				// Remove authorization scheme prefix (e.g. "Bearer")
				cred := strings.SplitN(*payload.BearerToken, " ", 2)[1]
				payload.BearerToken = &cred
			}
		}

		return payload, nil
	}
}

// EncodeGetFormationActivityError returns an encoder for errors returned by
// the get_formation_activity lfx_v2_formation_service endpoint.
func EncodeGetFormationActivityError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "NotFound":
			var res *lfxv2formationservice.NotFoundError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewGetFormationActivityNotFoundResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusNotFound)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeLivezResponse returns an encoder for responses returned by the
// lfx_v2_formation_service livez endpoint.
func EncodeLivezResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.([]byte)
		ctx = context.WithValue(ctx, goahttp.ContentTypeKey, "text/plain")
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeReadyzResponse returns an encoder for responses returned by the
// lfx_v2_formation_service readyz endpoint.
func EncodeReadyzResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.([]byte)
		ctx = context.WithValue(ctx, goahttp.ContentTypeKey, "text/plain")
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeReadyzError returns an encoder for errors returned by the readyz
// lfx_v2_formation_service endpoint.
func EncodeReadyzError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "ServiceUnavailable":
			var res *lfxv2formationservice.ServiceUnavailableError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewReadyzServiceUnavailableResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusServiceUnavailable)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// marshalLfxv2formationserviceviewsFormationSectionViewToFormationSectionResponseBody
// builds a value of type *FormationSectionResponseBody from a value of type
// *lfxv2formationserviceviews.FormationSectionView.
func marshalLfxv2formationserviceviewsFormationSectionViewToFormationSectionResponseBody(v *lfxv2formationserviceviews.FormationSectionView) *FormationSectionResponseBody {
	res := &FormationSectionResponseBody{
		Key:      *v.Key,
		Title:    *v.Title,
		Position: *v.Position,
	}

	return res
}

// marshalLfxv2formationserviceviewsFormationItemViewToFormationItemResponseBody
// builds a value of type *FormationItemResponseBody from a value of type
// *lfxv2formationserviceviews.FormationItemView.
func marshalLfxv2formationserviceviewsFormationItemViewToFormationItemResponseBody(v *lfxv2formationserviceviews.FormationItemView) *FormationItemResponseBody {
	res := &FormationItemResponseBody{
		UID:            *v.UID,
		ItemKey:        *v.ItemKey,
		SectionKey:     *v.SectionKey,
		Position:       *v.Position,
		Title:          *v.Title,
		OwnerTeam:      v.OwnerTeam,
		Gate:           *v.Gate,
		RequiresWriter: *v.RequiresWriter,
		StatusSource:   *v.StatusSource,
		ActionLink:     v.ActionLink,
		EvidenceLink:   v.EvidenceLink,
		Status:         *v.Status,
		Assignee:       v.Assignee,
		DueDate:        v.DueDate,
		Note:           v.Note,
		SkipReason:     v.SkipReason,
		Version:        *v.Version,
	}
	if v.PlatformCheck != nil {
		res.PlatformCheck = marshalLfxv2formationserviceviewsFormationPlatformCheckViewToFormationPlatformCheckResponseBody(v.PlatformCheck)
	}
	if v.ResolvedRef != nil {
		res.ResolvedRef = marshalLfxv2formationserviceviewsFormationResolvedRefViewToFormationResolvedRefResponseBody(v.ResolvedRef)
	}
	if v.SubItems != nil {
		res.SubItems = make([]*FormationSubItemResponseBody, len(v.SubItems))
		for i, val := range v.SubItems {
			if val == nil {
				res.SubItems[i] = nil
				continue
			}
			res.SubItems[i] = marshalLfxv2formationserviceviewsFormationSubItemViewToFormationSubItemResponseBody(val)
		}
	}

	return res
}

// marshalLfxv2formationserviceviewsFormationPlatformCheckViewToFormationPlatformCheckResponseBody
// builds a value of type *FormationPlatformCheckResponseBody from a value of
// type *lfxv2formationserviceviews.FormationPlatformCheckView.
func marshalLfxv2formationserviceviewsFormationPlatformCheckViewToFormationPlatformCheckResponseBody(v *lfxv2formationserviceviews.FormationPlatformCheckView) *FormationPlatformCheckResponseBody {
	if v == nil {
		return nil
	}
	res := &FormationPlatformCheckResponseBody{
		ResourceType: v.ResourceType,
		MinCount:     v.MinCount,
	}

	return res
}

// marshalLfxv2formationserviceviewsFormationResolvedRefViewToFormationResolvedRefResponseBody
// builds a value of type *FormationResolvedRefResponseBody from a value of
// type *lfxv2formationserviceviews.FormationResolvedRefView.
func marshalLfxv2formationserviceviewsFormationResolvedRefViewToFormationResolvedRefResponseBody(v *lfxv2formationserviceviews.FormationResolvedRefView) *FormationResolvedRefResponseBody {
	if v == nil {
		return nil
	}
	res := &FormationResolvedRefResponseBody{
		Type: v.Type,
		UID:  v.UID,
	}

	return res
}

// marshalLfxv2formationserviceviewsFormationSubItemViewToFormationSubItemResponseBody
// builds a value of type *FormationSubItemResponseBody from a value of type
// *lfxv2formationserviceviews.FormationSubItemView.
func marshalLfxv2formationserviceviewsFormationSubItemViewToFormationSubItemResponseBody(v *lfxv2formationserviceviews.FormationSubItemView) *FormationSubItemResponseBody {
	if v == nil {
		return nil
	}
	res := &FormationSubItemResponseBody{
		Key:    *v.Key,
		Title:  *v.Title,
		Status: *v.Status,
	}

	return res
}

// marshalLfxv2formationserviceviewsFormationProgressViewToFormationProgressResponseBody
// builds a value of type *FormationProgressResponseBody from a value of type
// *lfxv2formationserviceviews.FormationProgressView.
func marshalLfxv2formationserviceviewsFormationProgressViewToFormationProgressResponseBody(v *lfxv2formationserviceviews.FormationProgressView) *FormationProgressResponseBody {
	res := &FormationProgressResponseBody{
		NotStarted:         *v.NotStarted,
		InProgress:         *v.InProgress,
		Blocked:            *v.Blocked,
		AwaitingAcceptance: *v.AwaitingAcceptance,
		Done:               *v.Done,
		Skipped:            *v.Skipped,
	}

	return res
}

// marshalLfxv2formationserviceviewsFormationActivityEntryViewToFormationActivityEntryResponseBody
// builds a value of type *FormationActivityEntryResponseBody from a value of
// type *lfxv2formationserviceviews.FormationActivityEntryView.
func marshalLfxv2formationserviceviewsFormationActivityEntryViewToFormationActivityEntryResponseBody(v *lfxv2formationserviceviews.FormationActivityEntryView) *FormationActivityEntryResponseBody {
	res := &FormationActivityEntryResponseBody{
		Ulid:    *v.Ulid,
		ItemUID: v.ItemUID,
		Actor:   *v.Actor,
		SetBy:   *v.SetBy,
		Action:  *v.Action,
		Before:  v.Before,
		After:   v.After,
		At:      *v.At,
	}

	return res
}